Change forgenkey datatype

+++ database/migrations/2019_05_15_003627_create_shippings_table.php
@@ -15,7 +15,7 @@ class CreateShippingsTable extends Migration
     {
         Schema::create('shippings', function (Blueprint $table) {
             $table->bigIncrements('id');
-            $table->integer('customer_id')->nullable();
+            $table->unsignedTinyInteger('customer_id')->nullable();
             $table->string('name');
             $table->string('email');
             $table->string('phone');
